Add StableRecyclerViewAdapter, this closes #4
