+++ recycler/src/main/java/com/elpassion/android/commons/recycler/StableItemAdapter.kt
@@ -1,0 +1,7 @@
+package com.elpassion.android.commons.recycler
+
+import android.support.annotation.LayoutRes
+import android.support.v7.widget.RecyclerView
+
+abstract class StableItemAdapter<VH : RecyclerView.ViewHolder>(val stableId: Long,
+                                                               @LayoutRes layoutId: Int) : ItemAdapter<VH>(layoutId)
