+++ recycler/src/main/java/com/elpassion/android/commons/recycler/StableRecyclerViewAdapter.kt
@@ -1,0 +1,10 @@
+package com.elpassion.android.commons.recycler
+
+open class StableRecyclerViewAdapter : GenericRecyclerViewAdapter<StableItemAdapter<*>>(){
+
+    init {
+        setHasStableIds(true)
+    }
+
+    override fun getItemId(position: Int): Long = adapters[position].stableId
+}
